requirements.txt
absl-py==1.3.0
addict==2.4.0
altgraph==0.17.4
arrow==1.2.3
attrs==24.2.0
backcall==0.2.0
bayesian-optimization==1.0.1
binaryornot==0.4.4
boto3==1.4.4
botocore==1.5.95
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1.post20240308173724
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.3
configparser==5.3.0
cookiecutter==2.6.0
cryptography==3.4.7
cycler==0.11.0
Cython==0.27.3
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
docutils==0.19
easydict==1.9
ecdsa==0.19.0
entrypoints==0.4
ephemeral-port-reserve==1.1.4
esdk-obs-python==3.21.4
filelock==3.0.12
Flask==2.0.1
fonttools==4.38.0
future==0.18.2.post20200723173923
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gunicorn==20.1.0
huaweicloudsdkcore==3.1.106
huaweicloudsdkcsms==3.1.106
hyperopt==0.1.2
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.7.0
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.2.0
jupyter-client==7.4.4
jupyter-core==4.11.2
kiwisolver==1.4.4
lazy-import==0.2.2
lxml==4.9.1
ma-cli==1.2.3
ma-tensorboard @ file:///modelarts/tools/visualization/ma_tensorboard/ma_tensorboard-1.0.0-py3-none-any.whl
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv==1.2.7
mock==5.1.0
modelarts @ https://cnnorth4-modelarts-sdk.obs.cn-north-4.myhuaweicloud.com/modelarts-latest-py2.py3-none-any.whl
modelarts-pytorch-model-server @ http://100.95.151.167:6868/x86_64/ubuntu/dls-release/ubuntu-18.04/modelarts_service/modelarts_pytorch_model_server-1.0.4-py3-none-any.whl
moxing-framework @ http://100.95.151.167:6868/aarch64/euler/dls-release/ubuntu-16.04/moxing_framework/moxing_framework-2.1.0.5d9c87c8-py2.py3-none-any.whl
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.1.2.30
packaging==24.0
pandas==1.1.5
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prettytable==0.7.2
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.7
pymongo==4.3.2
pyparsing==3.0.9
pytest-runner==5.3.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2022.6
PyYAML==5.1
pyzmq==24.0.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.8.0
rsa==4.9
s3transfer==0.1.13
scikit-learn==0.22.1
scipy==1.5.2
semantic-version==2.10.0
sentencepiece==0.1.97
simplejson==3.19.3
six==1.16.0
sklearn==0.0
sortedcontainers==2.2.2
statistics==1.0.3.5
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.1.1
tensorboardX==2.0
text-unidecode==1.3
torch==1.8.0
torchtext==0.5.0
torchvision==0.9.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.12
watchdog==2.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.10.0
